refactor: extract theme helpers to avoid circular import
